analytics-service/requirements.txt
mysql-connector-python==8.0.33
pymongo==4.5.0 